Fix compatibility issue for Windows (Issue#1).
Replace "which" command with "where" for Windows instances. Fixes https://github.com/chriseppstein/compass-validator/issues/1.

"By default, the search is done along the current directory and in the paths specified by the PATH environment variable" (from "where" command-line help). In some cases, this fails to find a Java installation located on a different drive than the drive from which the validator is being called. To counter this, this code adds a second "where" call with the C:\ path hard-coded should the default search path(s) fail. This hard-coded path should eventually be converted to a configuration value to allow override.

Tested successfully on Windows Server 2012 R2 Standard.